feat: added gl_vec functions and header.
